requirements.txt
# Core build tools
setuptools>=65.0.0
wheel>=0.40.0
pip>=23.0

# Core data processing and numerical computing
numpy>=1.24.3,<2.0
pandas>=2.0.3,<3.0
scipy>=1.11.4

# ML/AI libraries
scikit-learn>=1.3.2
joblib>=1.3.2

# PyTorch (CPU version - compatible with numpy/pandas versions)
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cpu

# Transformers and NLP
transformers>=4.35.2
tokenizers>=0.14.1
safetensors>=0.4.1

# MLflow and related
mlflow>=2.14.0
cloudpickle>=3.0.0
pyyaml>=6.0.1
packaging>=23.2
protobuf>=5.0.0,<7.0.0

# Web framework
Flask>=2.2.5
Werkzeug>=2.2.3

# HTTP and networking
requests>=2.31.0
urllib3>=2.0.7

# Orchestration
apache-airflow>=2.7.3

# Configuration and environment
python-dotenv==1.0.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Additional utilities
python-dateutil>=2.8.2
typing-extensions>=4.7.1
